Final Images 	renamed:    src/components/Footer.css -> src/components/Footer/Footer.css 	renamed:    src/components/Footer.tsx -> src/components/Footer/Footer.tsx 	modified:   src/components/Landing_Page/Body.tsx 	modified:   src/pages/LandingPage.tsx
