+++ SDCTA/src/pages/LandingPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Body } from '../components/Landing_Page/Body';
-import {Footer} from '../components/Footer';
+import {Footer} from '../components/Footer/Footer';
 
 export const LandingPage = () => {
 	return (
